src/agent/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
durationpy==0.9
Flask==3.1.0
google-auth==2.38.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
kubernetes==32.0.1
MarkupSafe==3.0.2
mcp==1.6.0
oauthlib==3.2.2
priority==2.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
Quart==0.20.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
